Prevent stopping replaced chunk
A callback is registered to free trunk after it finished playing.
However, when we replace a trunk, we will manually free the trunk
and play a new trunk. However, freeing the old trunk will also
call the callback, and the callback will then try to free the new
trunk, which is unexpected.

We fixed it by playing the new trunk first and then manually free
the trunk, so that the channel will be in playing mode and thus
in the callback we can check if we still need to free the trunk
by checking if the channel is in playing mode.